Migrate interpreter byte storage to ByteVector
Completes the representation swap across local, ssh, agent, managed, worker,
cli and examples, so no public or internal surface still carries Vector[Byte].
ByteVector retains what its backing array retains rather than 2.4x it, and
reaches Array.clone parity at the array boundary every codec, digest and FS2
chunk conversion crosses, where Vector[Byte] was 6.8x slower.

Where a boundary already had a better route, take it: SshCapture and
CaptureState adopt chunk bytes directly, SshProcessRunner emits a byte chunk
instead of a boxed Seq, and LocalLogReader views its read buffer.

Pin the inline-script wire shape. encodeSubmitRequest routes an inline script's
bytes through circe's derived ScriptSource codec, but every existing golden used
ExistingRemote and so carried no bytes; nothing would have caught the derived
codec's numeric array turning into some other shape. Explicit ByteVector
instances hold that format and a new golden asserts the exact literal.

+++ modules/local/src/test/scala/io/github/bbuchsbaum/slurm4s/local/LocalCaptureBoundSuite.scala
@@ -97,7 +97,7 @@ class LocalCaptureBoundSuite extends munit.CatsEffectSuite:
               case other                             => fail(s"unexpected result: $other")
             assertEquals(
               evidence.bytes.size,
-              ByteLimit.defaultEvidence.value,
+              ByteLimit.defaultEvidence.value.toLong,
               "retained evidence must respect the evidence bound"
             )
             assert(evidence.truncated, "narrowed evidence must be reported as truncated")
